requirements.txt
Django==1.11.7
psycopg2==2.7.3.2
django-admin-sortable==2.1.1